# docs/public/api/openapi-3.0.yaml
openapi: 3.0.0
info:
  description: Track, Manage, and Organize your Things.
  title: Homebox API
  contact:
    name: Homebox Team
    url: https://discord.homebox.software
  version: "1.0"
paths:
  /v1/actions/create-missing-thumbnails:
    post:
      security:
        - Bearer: []
      description: Creates thumbnails for items that are missing them
      tags:
        - Actions
      summary: Create Missing Thumbnails
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v1.ActionAmountResult"
  /v1/actions/ensure-asset-ids:
    post:
      security:
        - Bearer: []
      description: Ensures all items in the database have an asset ID
      tags:
        - Actions
      summary: Ensure Asset IDs
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v1.ActionAmountResult"
  /v1/actions/ensure-import-refs:
    post:
      security:
        - Bearer: []
      description: Ensures all items in the database have an import ref
      tags:
        - Actions
      summary: Ensures Import Refs
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v1.ActionAmountResult"
  /v1/actions/set-primary-photos:
    post:
      security:
        - Bearer: []
      description: Sets the first photo of each item as the primary photo
      tags:
        - Actions
      summary: Set Primary Photos
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v1.ActionAmountResult"
  /v1/actions/wipe-inventory:
    post:
      security:
        - Bearer: []
      description: Deletes all items in the inventory
      tags:
        - Actions
      summary: Wipe Inventory
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/v1.WipeInventoryOptions"
        description: Wipe options
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v1.ActionAmountResult"
  /v1/actions/zero-item-time-fields:
    post:
      security:
        - Bearer: []
      description: Resets all item date fields to the beginning of the day
      tags:
        - Actions
      summary: Zero Out Time Fields
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v1.ActionAmountResult"
  "/v1/assets/{id}":
    get:
      security:
        - Bearer: []
      tags:
        - Items
      summary: Get Item by Asset ID
      parameters:
        - description: Asset ID
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/repo.PaginationResult-repo_EntitySummary"
...
  /v1/entities:
    get:
      security:
        - Bearer: []
      tags:
        - Entities
      summary: Query All Entities
      parameters:
        - description: search string
          name: q
          in: query
          schema:
            type: string
        - description: page number
          name: page
          in: query
          schema:
            type: integer
        - description: items per page
          name: pageSize
          in: query
          schema:
            type: integer
        - description: tags Ids
          name: tags
          in: query
          explode: true
          schema:
            type: array
            items:
              type: string
        - description: parent Ids
          name: parentIds
          in: query
          explode: true
          schema:
            type: array
            items:
              type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/repo.EntityListResult"
    post:
      security:
        - Bearer: []
      tags:
        - Entities
      summary: Create Entity
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/repo.EntityCreate"
        description: Entity Data
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/repo.EntityOut"
  /v1/entities/export:
    get:
      security:
        - Bearer: []
      tags:
        - Entities
      summary: Export Entities
      responses:
        "200":
          description: text/csv
          content:
            "*/*":
              schema:
                type: string
  /v1/entities/fields:
    get:
      security:
        - Bearer: []
      tags:
        - Entities
      summary: Get All Custom Field Names
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
  /v1/entities/fields/values:
    get:
      security:
        - Bearer: []
      tags:
        - Entities
      summary: Get All Custom Field Values
      parameters:
        - description: Field name
          name: field
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
  /v1/entities/import:
    post:
      security:
        - Bearer: []
      tags:
        - Entities
      summary: Import Entities
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                csv:
                  description: CSV file to upload
                  type: string
                  format: binary
              required:
                - csv
        required: true
      responses:
        "204":
          description: No Content
  /v1/entities/tree:
    get:
      security:
        - Bearer: []
      tags:
        - Entities
      summary: Get Locations Tree
      parameters:
        - description: include items in response tree
          name: withItems
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/repo.TreeItem"
  "/v1/entities/{id}":
    get:
      security:
        - Bearer: []
      tags:
        - Entities
      summary: Get Entity
      parameters:
        - description: Entity ID
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/repo.EntityOut"
    put:
      security:
        - Bearer: []
      tags:
        - Entities
      summary: Update Entity
      parameters:
        - description: Entity ID
          name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/repo.EntityUpdate"
        description: Entity Data
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/repo.EntityOut"
    delete:
      security:
        - Bearer: []
      tags:
        - Entities
      summary: Delete Entity
      parameters:
        - description: Entity ID
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No Content
    patch:
      security:
        - Bearer: []
      tags:
        - Entities
      summary: Patch Entity
      parameters:
        - description: Entity ID
          name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/repo.EntityPatch"
        description: Entity Data
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/repo.EntityOut"
  "/v1/entities/{id}/attachments":
    post:
      security:
        - Bearer: []
      tags:
        - Entities Attachments
      summary: Create Entity Attachment
      parameters:
        - description: Entity ID
          name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  description: File attachment
                  type: string
                  format: binary
                type:
                  description: Type of file
                  type: string
                primary:
                  description: Is this the primary attachment
                  type: boolean
                name:
                  description: name of the file including extension
                  type: string
              required:
                - file
                - name
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/repo.EntityOut"
        "422":
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/validate.ErrorResponse"
  "/v1/entities/{id}/attachments/external":
    post:
      security:
        - Bearer: []
      description: Links an entity to a document or URL in an external system without
        copying
      tags:
        - Entities Attachments
      summary: Create External Link Attachment
      parameters:
        - description: Entity ID
          name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/v1.externalAttachmentRequest"
        description: External document reference
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/repo.EntityOut"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/validate.ErrorResponse"
  "/v1/entities/{id}/attachments/{attachment_id}":
    get:
      security:
        - Bearer: []
      tags:
        - Entities Attachments
      summary: Get Entity Attachment
      parameters:
        - description: Entity ID
          name: id
          in: path
          required: true
          schema:
            type: string
        - description: Attachment ID
          name: attachment_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
    put:
      security:
        - Bearer: []
      tags:
        - Entities Attachments
      summary: Update Entity Attachment
      parameters:
        - description: Entity ID
          name: id
          in: path
          required: true
          schema:
            type: string
        - description: Attachment ID
          name: attachment_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/repo.ItemAttachmentUpdate"
        description: Attachment Update
        required: true
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/repo.EntityOut"
    delete:
      security:
        - Bearer: []
      tags:
        - Entities Attachments
      summary: Delete Entity Attachment
      parameters:
        - description: Entity ID
          name: id
          in: path
          required: true
          schema:
            type: string
        - description: Attachment ID
          name: attachment_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No Content
  "/v1/entities/{id}/duplicate":
    post:
      security:
        - Bearer: []
      tags:
        - Entities
      summary: Duplicate Entity
      parameters:
        - description: Entity ID
          name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/repo.DuplicateOptions"
        description: Duplicate Options
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/repo.EntityOut"
  "/v1/entities/{id}/layout":
    get:
      security:
        - Bearer: []
      tags:
        - Locations
      summary: Get a location overhead layout
      parameters:
        - description: Location ID
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/repo.LocationLayoutOut"
    put:
      security:
        - Bearer: []
      tags:
        - Locations
      summary: Replace a location overhead layout
      parameters:
        - description: Location ID
          name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/repo.LocationLayoutReplace"
        description: Complete layout
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/repo.LocationLayoutOut"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/validate.ErrorResponse"
    delete:
      security:
        - Bearer: []
      tags:
        - Locations
      summary: Delete a location overhead layout
      parameters:
        - description: Location ID
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No Content
  "/v1/entities/{id}/maintenance":
    get:
      security:
        - Bearer: []
      tags:
        - Item Maintenance
      summary: Get Maintenance Log
      parameters:
        - description: Item ID
          name: id
          in: path
          required: true
          schema:
            type: string
        - x-enum-varnames:
            - MaintenanceFilterStatusScheduled
            - MaintenanceFilterStatusCompleted
            - MaintenanceFilterStatusBoth
          name: status
          in: query
          schema:
            type: string
            enum:
              - scheduled
              - completed
              - both
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/repo.MaintenanceEntryWithDetails"
    post:
      security:
        - Bearer: []
      tags:
        - Item Maintenance
      summary: Create Maintenance Entry
      parameters:
        - description: Item ID
          name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/repo.MaintenanceEntryCreate"
        description: Entry Data
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/repo.MaintenanceEntry"
  "/v1/entities/{id}/path":
    get:
      security:
        - Bearer: []
      tags:
        - Entities
      summary: Get the full path of an entity
      parameters:
        - description: Entity ID
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/repo.EntityPath"
...
  /v1/users/logout:
    post:
      security:
        - Bearer: []
      tags:
        - Authentication
      summary: User Logout
      responses:
        "204":
          description: No Content
...
servers:
  - url: /api
components:
...
  securitySchemes:
    Bearer:
      description: "\"Type 'Bearer TOKEN' to correctly set the API Key\""
      type: apiKey
      name: Authorization
      in: header
  schemas:
...
    repo.EntityListResult:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/repo.EntitySummary"
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
        totalPrice:
          type: number
...
    repo.PaginationResult-repo_EntitySummary:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/repo.EntitySummary"
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
...
    v1.ActionAmountResult:
      type: object
      properties:
        completed:
          type: integer
...
    v1.WipeInventoryOptions:
      type: object
      properties:
        wipeLocations:
          type: boolean
        wipeMaintenance:
          type: boolean
        wipeTags:
          type: boolean
...
    validate.ErrorResponse:
      type: object
      properties:
        error:
          type: string
        fields:
          type: string